refactor: MovieList 컴포넌트 생성
