+++ src/App.tsx
@@ -1,13 +1,13 @@
-import React, { useState, useEffect, useMemo } from "react";
+import React, { useState, useEffect } from "react";
 import { themes, type ThemeName } from "./styles/theme";
 import type { Movie } from "./types/movie";
 import { Header } from "./components/Header";
 import { MovieForm } from "./components/MovieForm";
+import { MovieList } from "./components/MovieList";
 
 function App() {
   const [themeName, setThemeName] = useState<ThemeName>("light");
   const [movies, setMovies] = useState<Movie[]>([]);
-  const [searchTerm, setSearchTerm] = useState<string>("");
   const [error, setError] = useState<string | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
@@ -33,12 +33,6 @@ function App() {
     };
     fetchMovies();
   }, []);
-
-  const filteredMovies = useMemo(() => {
-    return movies.filter((movie) =>
-      movie.title.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-  }, [movies, searchTerm]);
 
   const handleAddMovie = async (newMovie: Omit<Movie, "id">) => {
     try {
@@ -100,58 +94,11 @@ function App() {
         onAddMovie={handleAddMovie}
         isLoading={isLoading}
       />
-      <div
-        style={{
-          padding: "20px",
-          borderRadius: "12px",
-          backgroundColor: currentTheme.componentBg,
-          border: `1px solid ${currentTheme.border}`,
-        }}
-      >
-        <h2>영화 목록</h2>
-        <input
-          type="text"
-          placeholder="검색..."
-          value={searchTerm}
-          onChange={(e) => setSearchTerm(e.target.value)}
-          style={{
-            padding: "8px",
-            borderRadius: "8px",
-            border: `1px solid ${currentTheme.border}`,
-            backgroundColor: currentTheme.inputBg,
-            color: currentTheme.text,
-            marginBottom: "16px",
-            width: "100%",
-          }}
-        />
-        {isLoading ? (
-          <div>로딩 중...</div>
-        ) : filteredMovies.length === 0 ? (
-          <div>영화가 없습니다</div>
-        ) : (
-          <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
-            {filteredMovies.map((movie) => (
-              <div
-                key={movie.id}
-                style={{
-                  padding: "12px",
-                  borderRadius: "8px",
-                  backgroundColor: currentTheme.hoverBg,
-                  display: "flex",
-                  flexDirection: "column",
-                  gap: "4px",
-                }}
-              >
-                <span>
-                  {movie.title} ({movie.year}) - {movie.director}
-                </span>
-                <span>장르: {movie.genre}</span>
-                <span>⭐: {movie.rating}</span>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
+      <MovieList 
+        movies={movies}
+        currentTheme={currentTheme}
+        isLoading={isLoading}
+      />
     </div>
   );
 }
